+++ lib/aspera/cli/plugins/orchestrator.rb
@@ -120,49 +120,49 @@ module Aspera
             )
           end)
           command :status,     description: 'Check running status of workflows',
-            arguments: [{name: :wf_id, type: :identifier}],
-            action: ->(wf_id:, **) { Result::ObjectList.new(call_ao(wf_id.eql?(SpecialValues::ALL) ? 'workflows_status' : "workflows_status/#{wf_id}")['workflows']['workflow']) }
+            arguments: [{name: :workflow_id, type: :identifier}],
+            action: ->(workflow_id:, **) { Result::ObjectList.new(call_ao(workflow_id.eql?(SpecialValues::ALL) ? 'workflows_status' : "workflows_status/#{workflow_id}")['workflows']['workflow']) }
           command :inputs,     description: 'Fetch input specification for a workflow',
-            arguments: [{name: :wf_id, type: :identifier}],
-            action: ->(wf_id:, **) { Result::SingleObject.new(call_ao("workflow_inputs_spec/#{wf_id}")['workflow_inputs_spec']) }
+            arguments: [{name: :workflow_id, type: :identifier}],
+            action: ->(workflow_id:, **) { Result::SingleObject.new(call_ao("workflow_inputs_spec/#{workflow_id}")['workflow_inputs_spec']) }
           command :details,    description: 'Check detailed running status of a workflow',
-            arguments: [{name: :wf_id, type: :identifier}],
-            action: ->(wf_id:, **) { Result::ObjectList.new(call_ao("workflow_details/#{wf_id}")['workflows']['workflow']['statuses']) }
+            arguments: [{name: :workflow_id, type: :identifier}],
+            action: ->(workflow_id:, **) { Result::ObjectList.new(call_ao("workflow_details/#{workflow_id}")['workflows']['workflow']['statuses']) }
           command :start,      description: 'Initiate a work order (sync or async)',
-            arguments: [{name: :wf_id, type: :identifier}, {name: :parameters, type: Hash, mandatory: false, default: {}}]
+            arguments: [{name: :workflow_id, type: :identifier}, {name: :parameters, type: Hash, mandatory: false, default: {}}]
           command :export,     description: 'Export a workflow',
-            arguments: [{name: :wf_id, type: :identifier}],
-            action: ->(wf_id:, **) { Result::Text.new(call_ao("export_workflow/#{wf_id}", format: nil, http: true).body) }
+            arguments: [{name: :workflow_id, type: :identifier}],
+            action: ->(workflow_id:, **) { Result::Text.new(call_ao("export_workflow/#{workflow_id}", format: nil, http: true).body) }
           command :workorders, description: 'Fetch all work orders from a workflow',
-            arguments: [{name: :wf_id, type: :identifier}],
-            action: ->(wf_id:, **) { Result::ObjectList.new(call_ao("work_orders_list/#{wf_id}")['work_orders']) }
+            arguments: [{name: :workflow_id, type: :identifier}],
+            action: ->(workflow_id:, **) { Result::ObjectList.new(call_ao("work_orders_list/#{workflow_id}")['work_orders']) }
           command :outputs,    description: 'Fetch output specification for a workflow',
-            arguments: [{name: :wf_id, type: :identifier}],
-            action: ->(wf_id:, **) { Result::ObjectList.new(call_ao("workflow_outputs_spec/#{wf_id}")['workflow_outputs_spec']['output']) }
+            arguments: [{name: :workflow_id, type: :identifier}],
+            action: ->(workflow_id:, **) { Result::ObjectList.new(call_ao("workflow_outputs_spec/#{workflow_id}")['workflow_outputs_spec']['output']) }
         end
 
         commands_under :workorders do
           command :status, description: 'Check the status of a work order',
-            arguments: [{name: :wo_id, type: :identifier}],
-            action: ->(wo_id:, **) { Result::SingleObject.new(call_ao("work_order_status/#{wo_id}")['work_order']) }
+            arguments: [{name: :workorder_id, type: :identifier}],
+            action: ->(workorder_id:, **) { Result::SingleObject.new(call_ao("work_order_status/#{workorder_id}")['work_order']) }
           command :cancel, description: 'Cancel a work order',
-            arguments: [{name: :wo_id, type: :identifier}],
-            action: ->(wo_id:, **) { Result::SingleObject.new(call_ao("work_order_cancel/#{wo_id}")['work_order']) }
+            arguments: [{name: :workorder_id, type: :identifier}],
+            action: ->(workorder_id:, **) { Result::SingleObject.new(call_ao("work_order_cancel/#{workorder_id}")['work_order']) }
           command :reset,  description: 'Reset a work order',
-            arguments: [{name: :wo_id, type: :identifier}],
-            action: ->(wo_id:, **) { Result::SingleObject.new(call_ao("work_order_reset/#{wo_id}")['work_order']) }
+            arguments: [{name: :workorder_id, type: :identifier}],
+            action: ->(workorder_id:, **) { Result::SingleObject.new(call_ao("work_order_reset/#{workorder_id}")['work_order']) }
           command :output, description: 'Fetch output of a work order',
-            arguments: [{name: :wo_id, type: :identifier}],
-            action: ->(wo_id:, **) { Result::ObjectList.new(call_ao("work_order_output/#{wo_id}", format: 'xml')['variable']) }
+            arguments: [{name: :workorder_id, type: :identifier}],
+            action: ->(workorder_id:, **) { Result::ObjectList.new(call_ao("work_order_output/#{workorder_id}", format: 'xml')['variable']) }
         end
 
         commands_under :workstep do
           command :status, description: 'Check the status of a work step',
-            arguments: [{name: :ws_id, type: :identifier}],
-            action: ->(ws_id:, **) { Result::SingleObject.new(call_ao("work_step_status/#{ws_id}")) }
+            arguments: [{name: :workstep_id, type: :identifier}],
+            action: ->(workstep_id:, **) { Result::SingleObject.new(call_ao("work_step_status/#{workstep_id}")) }
           command :cancel, description: 'Cancel a work step',
-            arguments: [{name: :ws_id, type: :identifier}],
-            action: ->(ws_id:, **) { Result::SingleObject.new(call_ao("work_step_cancel/#{ws_id}")) }
+            arguments: [{name: :workstep_id, type: :identifier}],
+            action: ->(workstep_id:, **) { Result::SingleObject.new(call_ao("work_step_cancel/#{workstep_id}")) }
         end
 
         # --- setup ---
@@ -209,7 +209,7 @@ module Aspera
         end
 
         # 2.1/2.2 Initiate a workorder (async / synchronous)
-        def action_workflows_start(wf_id:, parameters:, **)
+        def action_workflows_start(workflow_id:, parameters:, **)
           call_params = {format: :json}
           # get external parameters if any
           parameters.each do |name, value|
@@ -227,7 +227,7 @@ module Aspera
             # implicitly, call is synchronous
             call_params['synchronous'] = true
           end
-          result_data = call_ao("initiate/#{wf_id}", args: call_params)
+          result_data = call_ao("initiate/#{workflow_id}", args: call_params)
           call_params['synchronous'] ? Result::Text.new(result_data) : Result::SingleObject.new(result_data)
         end
       end